moved incorrect service to helper, and updated test service
